backend/requirements.txt
Werkzeug==3.0.3
python-dotenv==1.0.1
groq==0.5.0
Flask==3.0.3
psycopg2-binary==2.9.9
Flask-Cors==4.0.1
requests==2.31.0
